part 1 safety razor upload
(uploaded part 1 of the safety razor. Includes design for the top template through following the Learn It! tutorial video. Includes the drawing as well as the 3d design. Linked youtube video that was used to complete the build is in the description.) https://www.youtube.com/watch?v=nuchVR29v0o